LLM/STT refactor
